+++ migrations/20221030120533_add-general-topic.js
@@ -1,0 +1,8 @@
+
+exports.up = async function (knex) {
+  await knex.raw(`insert into tags (name) values ('general') ON CONFLICT DO NOTHING`)
+}
+
+exports.down = function (knex) {
+  
+}